refactor: localStorage를 store가 아닌 localStorage 폴더에서 관리
